@@ -4,9 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
@@ -15,25 +15,25 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SessionDto {
-    private Long id;
+	private Long id;
 
-    @NotBlank
-    @Size(max = 50)
-    private String name;
+	@NotBlank
+	@Size(max = 50)
+	private String name;
 
-    @NotNull
-    private Date date;
+	@NotNull
+	private Date date;
 
-    @NotNull
-    private Long teacher_id;
+	@NotNull
+	private Long teacher_id;
 
-    @NotNull
-    @Size(max = 2500)
-    private String description;
+	@NotNull
+	@Size(max = 2500)
+	private String description;
 
-    private List<Long> users;
+	private List<Long> users;
 
-    private LocalDateTime createdAt;
+	private LocalDateTime createdAt;
 
-    private LocalDateTime updatedAt;
+	private LocalDateTime updatedAt;
 }
