@@ -4,24 +4,24 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class TeacherDto {
-    private Long id;
-    @NotBlank
-    @Size(max = 20)
-    private String lastName;
+	private Long id;
+	@NotBlank
+	@Size(max = 20)
+	private String lastName;
 
-    @NotBlank
-    @Size(max = 20)
-    private String firstName;
+	@NotBlank
+	@Size(max = 20)
+	private String firstName;
 
-    private LocalDateTime createdAt;
+	private LocalDateTime createdAt;
 
-    private LocalDateTime updatedAt;
+	private LocalDateTime updatedAt;
 }
