@@ -11,26 +11,27 @@ import com.openclassrooms.starterjwt.repository.UserRepository;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
-  UserRepository userRepository;
 
-  UserDetailsServiceImpl(UserRepository userRepository) {
-    this.userRepository = userRepository;
-  }
+	UserRepository userRepository;
 
-  @Override
-  @Transactional
-  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-    User user = userRepository.findByEmail(username)
-        .orElseThrow(() -> new UsernameNotFoundException("User Not Found with email: " + username));
+	UserDetailsServiceImpl(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
 
-    return UserDetailsImpl
-            .builder()
-            .id(user.getId())
-            .username(user.getEmail())
-            .lastName(user.getLastName())
-            .firstName(user.getFirstName())
-            .password(user.getPassword())
-            .build();
-  }
+	@Override
+	@Transactional
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		User user = userRepository.findByEmail(username)
+				.orElseThrow(() -> new UsernameNotFoundException("User Not Found with email: " + username));
+
+		return UserDetailsImpl
+				.builder()
+				.id(user.getId())
+				.username(user.getEmail())
+				.lastName(user.getLastName())
+				.firstName(user.getFirstName())
+				.password(user.getPassword())
+				.build();
+	}
 
 }
